+++ packages/office-ui-fabric-react/src/components/Panel/Panel.doc.tsx
@@ -14,6 +14,7 @@ import { PanelLightDismissExample } from './examples/Panel.LightDismiss.Example'
 import { PanelLightDismissCustomExample } from './examples/Panel.LightDismissCustom.Example';
 import { PanelNonModalExample } from './examples/Panel.NonModal.Example';
 import { PanelFooterExample } from './examples/Panel.Footer.Example';
+import { PanelPreventDefaultExample } from './examples/Panel.PreventDefault.Example';
 import { PanelStatus } from './Panel.checklist';
 
 const PanelSmallRightExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Panel/examples/Panel.SmallRight.Example.tsx') as string;
@@ -30,6 +31,7 @@ const PanelLightDismissExampleCode = require('!raw-loader!office-ui-fabric-react
 const PanelLightDismissCustomExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Panel/examples/Panel.LightDismissCustom.Example.tsx') as string;
 const PanelNonModalExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Panel/examples/Panel.NonModal.Example.tsx') as string;
 const PanelFooterExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Panel/examples/Panel.Footer.Example.tsx') as string;
+const PanelPreventDefaultExampleCode = require('!raw-loader!office-ui-fabric-react/src/components/Panel/examples/Panel.PreventDefault.Example.tsx') as string;
 
 export const PanelPageProps: IDocPageProps = {
   title: 'Panel',
@@ -103,6 +105,11 @@ export const PanelPageProps: IDocPageProps = {
       title: 'Panel - Footer',
       code: PanelFooterExampleCode,
       view: <PanelFooterExample />
+    },
+    {
+      title: 'Panel - Prevent Default Sample',
+      code: PanelPreventDefaultExampleCode,
+      view: <PanelPreventDefaultExample />
     }
   ],
   propertiesTablesSources: [require<string>('!raw-loader!office-ui-fabric-react/src/components/Panel/Panel.types.ts')],